feat: add integrity to asset-manifest.json
BREAKING CHANGE: asset-manifest.json now contains an object with src and
integrity for each asset instead of previous src string.

Previous: { "asset.js": "asset.hash.js" }
New: { "asset.js": { src: "asset.hash.js", integrity: "sha256..." }
