+++ examples/basic/src/server/components/Html.jsx
@@ -2,6 +2,11 @@ import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import serialize from 'serialize-javascript';
 
+// Helpers
+function isAssetObject(asset) {
+  return typeof asset === 'object';
+}
+
 class Html extends PureComponent {
   render() {
     const {
@@ -29,20 +34,39 @@ class Html extends PureComponent {
           {meta}
           {noscript}
           {base}
-          {manifest && <link rel="manifest" href={manifest} />}
+          {manifest && isAssetObject(manifest) ? (
+            <link rel="manifest" href={manifest.src} integrity={manifest.integrity} crossOrigin="anonymous" />
+          ) : (
+            <link rel="manifest" href={manifest} crossOrigin="anonymous" />
+          )}
           {link}
-          {preload.map(file => {
-            if (/\.js$/.test(file)) return <link key={file} href={file} rel="preload" as="script" />;
-            if (/\.css$/.test(file)) return <link key={file} href={file} rel="preload" as="style" />;
-            if (/(\.woff|\.woff2|\.eot|\.ttf)$/.test(file))
-              return <link key={file} href={file} rel="preload" as="font" crossOrigin="anonymous" />;
-            if (/(\.png|\.jpe?g|\.gif)$/.test(file))
-              return <link key={file} href={file} rel="preload" as="image" crossOrigin="anonymous" />;
+          {preload.map(asset => {
+            const { src, integrity = undefined } = isAssetObject(asset) ? asset : { src: asset };
+            if (/\.js$/.test(src))
+              return (
+                <link key={src} href={src} rel="preload" as="script" integrity={integrity} crossOrigin="anonymous" />
+              );
+            if (/\.css$/.test(src))
+              return (
+                <link key={src} href={src} rel="preload" as="style" integrity={integrity} crossOrigin="anonymous" />
+              );
+            if (/(\.woff|\.woff2|\.eot|\.ttf)$/.test(src))
+              return (
+                <link key={src} href={src} rel="preload" as="font" integrity={integrity} crossOrigin="anonymous" />
+              );
+            if (/(\.png|\.jpe?g|\.gif)$/.test(src))
+              return (
+                <link key={src} href={src} rel="preload" as="image" integrity={integrity} crossOrigin="anonymous" />
+              );
             return null;
           })}
-          {bundleCss.map(css => (
-            <link key={css} href={css} rel="stylesheet" />
-          ))}
+          {bundleCss.map(css =>
+            isAssetObject(css) ? (
+              <link key={css.src} integrity={css.integrity} href={css.src} rel="stylesheet" crossOrigin="anonymous" />
+            ) : (
+              <link key={css} href={css} rel="stylesheet" crossOrigin="anonymous" />
+            ),
+          )}
           {style}
           {script}
         </head>
@@ -60,9 +84,20 @@ class Html extends PureComponent {
               }}
             />
           )}
-          {bundleJs.map(js => (
-            <script key={js} type="text/javascript" src={js} defer />
-          ))}
+          {bundleJs.map(js =>
+            isAssetObject(js) ? (
+              <script
+                key={js.src}
+                type="text/javascript"
+                src={js.src}
+                integrity={js.integrity}
+                defer
+                crossOrigin="anonymous"
+              />
+            ) : (
+              <script key={js} type="text/javascript" src={js} defer crossOrigin="anonymous" />
+            ),
+          )}
         </body>
       </html>
     );
@@ -78,12 +113,30 @@ Html.propTypes = {
   base: PropTypes.node,
   htmlAttributes: PropTypes.objectOf(PropTypes.shape({})),
   bodyAttributes: PropTypes.objectOf(PropTypes.shape({})),
-  bundleJs: PropTypes.arrayOf(PropTypes.string.isRequired),
-  bundleCss: PropTypes.arrayOf(PropTypes.string.isRequired),
-  manifest: PropTypes.string,
+  bundleJs: PropTypes.arrayOf(
+    PropTypes.oneOfType([
+      PropTypes.objectOf(PropTypes.shape({ src: PropTypes.string.isRequired, integrity: PropTypes.string })),
+      PropTypes.string.isRequired,
+    ]),
+  ),
+  bundleCss: PropTypes.arrayOf(
+    PropTypes.oneOfType([
+      PropTypes.objectOf(PropTypes.shape({ src: PropTypes.string.isRequired, integrity: PropTypes.string })),
+      PropTypes.string.isRequired,
+    ]),
+  ),
+  manifest: PropTypes.oneOfType([
+    PropTypes.objectOf(PropTypes.shape({ src: PropTypes.string.isRequired, integrity: PropTypes.string })),
+    PropTypes.string,
+  ]),
   body: PropTypes.string,
   initialState: PropTypes.objectOf(PropTypes.shape({})),
-  preload: PropTypes.arrayOf(PropTypes.string.isRequired),
+  preload: PropTypes.arrayOf(
+    PropTypes.oneOfType([
+      PropTypes.objectOf(PropTypes.shape({ src: PropTypes.string.isRequired, integrity: PropTypes.string })),
+      PropTypes.string.isRequired,
+    ]),
+  ),
 };
 Html.defaultProps = {
   title: null,
