+++ package.json
@@ -60,14 +60,14 @@
   },
   "dependencies": {},
   "devDependencies": {
-    "@commitlint/cli": "^8.1.0",
-    "@commitlint/config-conventional": "^8.1.0",
-    "@commitlint/config-lerna-scopes": "^8.1.0",
+    "@commitlint/cli": "^8.2.0",
+    "@commitlint/config-conventional": "^8.2.0",
+    "@commitlint/config-lerna-scopes": "^8.2.0",
     "@types/fs-extra": "^8.0.0",
     "@types/jest": "^24.0.15",
     "@types/node": "^12.6.9",
-    "@typescript-eslint/eslint-plugin": "^2.0.0",
-    "@typescript-eslint/parser": "^2.0.0",
+    "@typescript-eslint/eslint-plugin": "^2.3.0",
+    "@typescript-eslint/parser": "^2.3.0",
     "del-cli": "^3.0.0",
     "eslint": "^6.0.0",
     "eslint-config-airbnb": "^18.0.1",
@@ -81,7 +81,7 @@
     "husky": "^3.0.0",
     "jest": "^24.8.0",
     "lerna": "^3.15.0",
-    "lint-staged": "^9.2.1",
+    "lint-staged": "^9.3.0",
     "npm-run-all": "^4.1.5",
     "prettier": "^1.18.2",
     "ts-jest": "^24.0.2",
