+++ packages/babel-preset-varan/package.json
@@ -50,7 +50,7 @@
     "@babel/preset-react": "^7.0.0",
     "@babel/preset-typescript": "^7.3.3",
     "@svgr/webpack": "^4.3.2",
-    "babel-plugin-named-asset-import": "^0.3.2",
+    "babel-plugin-named-asset-import": "^0.3.4",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24"
   },
   "peerDependencies": {
