+++ packages/create-varan-app/package.json
@@ -55,7 +55,7 @@
     "listr": "^0.14.2",
     "lodash.defaults": "^4.2.0",
     "source-map-support": "^0.5.13",
-    "tar": "^4.4.10",
+    "tar": "^4.4.11",
     "update-notifier": "^3.0.1",
     "validate-npm-package-name": "^3.0.0",
     "write-pkg": "^4.0.0"
