+++ packages/varan/package.json
@@ -68,11 +68,11 @@
     "listr": "^0.14.2",
     "lodash": "^4.17.15",
     "mini-css-extract-plugin": "^0.8.0",
-    "ora": "^3.4.0",
+    "ora": "^4.0.0",
     "postcss-loader": "^3.0.0",
     "postcss-preset-env": "^6.7.0",
-    "react-app-polyfill": "^1.0.1",
-    "react-dev-utils": "^9.0.1",
+    "react-app-polyfill": "^1.0.3",
+    "react-dev-utils": "^9.0.4",
     "resolve-url-loader": "^3.1.0",
     "responsive-loader": "^1.2.0",
     "rxjs": "^6.5.2",
@@ -83,36 +83,47 @@
     "style-loader": "^1.0.0",
     "sw-precache-webpack-plugin": "^0.11.5",
     "table": "^5.4.5",
-    "terser-webpack-plugin": "^2.0.0",
+    "terser-webpack-plugin": "^2.1.0",
     "update-notifier": "^3.0.1",
     "url-loader": "^2.1.0",
     "wait-on": "^3.3.0",
     "webpack": "^4.39.1",
     "webpack-assets-manifest": "^3.1.1",
-    "webpack-bundle-analyzer": "^3.4.1",
-    "webpack-dev-server": "^3.7.2",
+    "webpack-bundle-analyzer": "^3.5.1",
+    "webpack-dev-server": "^3.8.1",
     "webpack-hot-client": "^4.1.2",
     "webpack-merge": "^4.2.1",
     "webpack-node-externals": "^1.7.2",
     "webpack-pwa-manifest": "^4.0.0",
+    "webpack-sources": "^1.4.3",
     "webpack-stats-plugin": "^0.3.0"
   },
   "devDependencies": {
     "@babel/runtime": "^7.5.5",
+    "@types/compression-webpack-plugin": "^2.0.1",
+    "@types/cssnano": "^4.0.0",
     "@types/express": "^4.17.0",
-    "@types/filesize": "^4.1.0",
+    "@types/filesize": "^4.2.0",
     "@types/fs-extra": "^8.0.0",
     "@types/listr": "^0.14.2",
     "@types/lodash": "^4.14.136",
     "@types/memory-fs": "^0.3.2",
+    "@types/mini-css-extract-plugin": "^0.8.0",
     "@types/react": "^16.8.24",
+    "@types/react-dev-utils": "^9.0.1",
     "@types/react-dom": "^16.8.5",
     "@types/split": "^1.0.0",
+    "@types/sw-precache-webpack-plugin": "^0.11.0",
     "@types/table": "^4.0.6",
+    "@types/terser-webpack-plugin": "^1.2.1",
     "@types/update-notifier": "^2.5.0",
+    "@types/webpack-assets-manifest": "^3.0.0",
+    "@types/webpack-bundle-analyzer": "^2.13.3",
     "@types/webpack-dev-server": "^3.1.7",
     "@types/webpack-merge": "^4.1.5",
-    "@typescript-eslint/eslint-plugin": "^2.0.0",
+    "@types/webpack-node-externals": "^1.6.3",
+    "@types/webpack-sources": "^0.1.5",
+    "@typescript-eslint/eslint-plugin": "^2.3.0",
     "babel-core": "^7.0.0-bridge.0",
     "express": "^4.17.1",
     "memory-fs": "^0.4.1",
