+++ packages/varan/src/index.ts
@@ -1,5 +1,8 @@
 import build from './lib/build';
 import watch from './lib/watch';
+import createCommonConfig from './webpack/createCommonConfig';
+import createClientConfig from './webpack/createClientConfig';
+import createServerConfig from './webpack/createServerConfig';
 
 // Exports
-export { build, watch };
+export { build, watch, createCommonConfig, createClientConfig, createServerConfig };
