+++ packages/varan/types/postcss-preset-env.d.ts
@@ -1,0 +1,6 @@
+// Type definitions for postcss-preset-env 6.7.0
+// TypeScript Version: 3.6.3
+
+declare module 'postcss-preset-env' {
+  export default function createPostCssPresetEnv(): {};
+}
